+++ ndk-build/src/manifest.rs
@@ -1,258 +1,288 @@
 use crate::error::NdkError;
-use std::fs::File;
-use std::io::Write;
-use std::path::Path;
+use serde::{Deserialize, Serialize, Serializer};
+use std::{fs::File, io::Write, path::Path};
 
-#[derive(Debug)]
-pub struct Manifest {
+#[serde(rename(serialize = "manifest"))]
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct AndroidManifest {
+    #[serde(skip_deserializing)]
+    #[serde(rename(serialize = "xmlns:android"))]
+    #[serde(default = "default_android_namespace")]
+    pub android_namespace: String,
+    #[serde(skip_deserializing)]
+    #[serde(rename(serialize = "package"))]
     pub package_name: String,
-    pub package_label: String,
-    pub version_name: String,
+    #[serde(skip_deserializing)]
+    #[serde(rename(serialize = "android:versionCode"))]
     pub version_code: u32,
-    pub target_name: String,
-    pub target_sdk_version: u32,
-    pub min_sdk_version: u32,
-    pub opengles_version: (u8, u8),
+    #[serde(skip_deserializing)]
+    #[serde(rename(serialize = "android:versionName"))]
+    pub version_name: String,
+
+    #[serde(rename(serialize = "uses-sdk"))]
+    #[serde(default)]
+    pub sdk: Sdk,
+    #[serde(rename(serialize = "uses-feature", deserialize = "feature"))]
+    #[serde(default)]
     pub features: Vec<Feature>,
-    pub permissions: Vec<Permission>,
-    pub intent_filters: Vec<IntentFilter>,
-    pub icon: Option<String>,
-    pub fullscreen: bool,
-    pub orientation: Option<String>,
-    pub launch_mode: Option<String>,
-    pub debuggable: bool,
-    pub split: Option<String>,
-    pub application_metadatas: Vec<ApplicationMetadata>,
-    pub activity_metadatas: Vec<ActivityMetadata>,
+    #[serde(rename(serialize = "uses-permission", deserialize = "permission"))]
+    pub permissions: Option<Vec<Permission>>,
+
+    #[serde(default)]
+    pub application: Application,
 }
 
-impl Manifest {
-    pub fn to_string(&self) -> String {
-        let split = if let Some(split) = self.split.as_ref() {
-            format!(r#"split="{}" android:isFeatureSplit="true""#, split)
-        } else {
-            "".to_string()
-        };
-        let (major, minor) = self.opengles_version;
-        let opengles_version = format!("0x{:04}{:04}", major, minor);
-
-        let icon = self
-            .icon
-            .as_ref()
-            .map(|icon| format!(r#"android:icon="{}""#, icon))
-            .unwrap_or_default();
-
-        let fullscreen = if self.fullscreen {
-            r#"android:theme="@android:style/Theme.DeviceDefault.NoActionBar.Fullscreen""#
-        } else {
-            ""
-        };
-
-        let orientation = self.orientation.as_deref().unwrap_or("unspecified");
-        let launch_mode = self.launch_mode.as_deref().unwrap_or("standard");
-
-        let features: Vec<String> = self.features.iter().map(|f| f.to_string()).collect();
-        let permissions: Vec<String> = self.permissions.iter().map(|p| p.to_string()).collect();
-        let intent_filters: Vec<String> =
-            self.intent_filters.iter().map(|i| i.to_string()).collect();
-        let application_metadatas: Vec<String> = self
-            .application_metadatas
-            .iter()
-            .map(|f| f.to_string())
-            .collect();
-        let activity_metadatas: Vec<String> = self
-            .activity_metadatas
-            .iter()
-            .map(|f| f.to_string())
-            .collect();
-
-        format!(
-            r#"<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="{package_name}"
-        android:versionCode="{version_code}"
-        android:versionName="{version_name}"
-        {split}>
-    <uses-sdk
-        android:targetSdkVersion="{target_sdk_version}"
-        android:minSdkVersion="{min_sdk_version}" />
-    <uses-feature android:glEsVersion="{opengles_version}" android:required="true"></uses-feature>
-    {features}
-    {permissions}
-    <application
-            android:hasCode="false"
-            android:label="{package_label}"
-            android:debuggable="{debuggable}"
-            {icon}
-            {fullscreen}>
-            {application_metadatas}
-        <activity
-                android:name="android.app.NativeActivity"
-                android:label="{package_label}"
-                android:screenOrientation="{orientation}"
-                android:launchMode="{launch_mode}"
-                android:configChanges="orientation|keyboardHidden|screenSize">
-            <meta-data android:name="android.app.lib_name" android:value="{target_name}" />
-            {activity_metadatas}
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-            {intent_filters}
-        </activity>
-    </application>
-</manifest>"#,
-            package_name = &self.package_name,
-            package_label = &self.package_label,
-            version_name = &self.version_name,
-            version_code = self.version_code,
-            split = split,
-            target_sdk_version = self.target_sdk_version,
-            min_sdk_version = self.min_sdk_version,
-            opengles_version = opengles_version,
-            target_name = &self.target_name,
-            icon = icon,
-            fullscreen = fullscreen,
-            orientation = orientation,
-            launch_mode = launch_mode,
-            application_metadatas = application_metadatas.join("\n"),
-            activity_metadatas = activity_metadatas.join("\n"),
-            debuggable = self.debuggable,
-            features = features.join("\n"),
-            permissions = permissions.join("\n"),
-            intent_filters = intent_filters.join("\n"),
-        )
-    }
-
+impl AndroidManifest {
     pub fn write_to(&self, dir: &Path) -> Result<(), NdkError> {
         let mut file = File::create(dir.join("AndroidManifest.xml"))?;
-        writeln!(file, "{}", self.to_string())?;
+        let serialized = quick_xml::se::to_string(&self).unwrap();
+        writeln!(file, r#"<?xml version="1.0" encoding="utf-8"?>"#)?;
+        writeln!(file, "{}", serialized)?;
         Ok(())
     }
 }
 
-#[derive(Debug)]
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct Application {
+    #[serde(rename(serialize = "android:debuggable"))]
+    pub debuggable: Option<bool>,
+    #[serde(rename(serialize = "android:theme"))]
+    #[serde(serialize_with = "serialize_fullscreen")]
+    pub fullscreen: Option<bool>,
+    #[serde(rename(serialize = "android:hasCode"))]
+    #[serde(default = "default_has_code")]
+    pub has_code: Option<bool>,
+    #[serde(rename(serialize = "android:icon"))]
+    pub icon: Option<String>,
+    #[serde(rename(serialize = "android:label"))]
+    #[serde(default)]
+    pub label: String,
+
+    #[serde(rename(serialize = "meta-data", deserialize = "metadata"))]
+    pub meta_datas: Option<Vec<MetaData>>,
+    #[serde(default)]
+    pub activity: Activity,
+}
+
+impl Default for Application {
+    fn default() -> Self {
+        Self {
+            debuggable: None,
+            has_code: default_has_code(),
+            icon: None,
+            label: "".to_string(),
+            fullscreen: None,
+            meta_datas: None,
+            activity: Activity::default(),
+        }
+    }
+}
+
+fn serialize_fullscreen<S>(v: &Option<bool>, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    match v {
+        Some(_) => {
+            serializer.serialize_some("@android:style/Theme.DeviceDefault.NoActionBar.Fullscreen")
+        }
+        None => serializer.serialize_none(),
+    }
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct Activity {
+    #[serde(rename(serialize = "android:configChanges"))]
+    #[serde(default = "default_config_changes")]
+    pub config_changes: Option<String>,
+    #[serde(rename(serialize = "android:label"))]
+    pub label: Option<String>,
+    #[serde(rename(serialize = "android:launchMode"))]
+    pub launch_mode: Option<String>,
+    #[serde(rename(serialize = "android:name"))]
+    #[serde(default = "default_activity_name")]
+    #[serde(skip_deserializing)]
+    pub name: Option<String>,
+    #[serde(rename(serialize = "android:screenOrientation"))]
+    pub orientation: Option<String>,
+
+    #[serde(rename(serialize = "meta-data", deserialize = "metadata"))]
+    #[serde(default)]
+    pub meta_datas: Vec<MetaData>,
+    #[serde(rename(serialize = "intent-filter", deserialize = "intent_filter"))]
+    #[serde(default)]
+    pub intent_filters: Vec<IntentFilter>,
+}
+
+impl Default for Activity {
+    fn default() -> Self {
+        Self {
+            name: default_activity_name(),
+            label: None,
+            orientation: None,
+            launch_mode: None,
+            config_changes: default_config_changes(),
+            meta_datas: Vec::new(),
+            intent_filters: Vec::new(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct IntentFilter {
+    // Serialize as struct for proper xml format
+    #[serde(serialize_with = "serialize_actions")]
+    #[serde(rename(serialize = "action"))]
+    pub actions: Vec<String>,
+    // Serialize as vector of structs for proper xml format
+    #[serde(serialize_with = "serialize_catergories")]
+    #[serde(rename(serialize = "category"))]
+    pub categories: Option<Vec<String>>,
+    pub data: Option<Vec<IntentFilterData>>,
+}
+
+fn serialize_actions<S>(actions: &Vec<String>, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    use serde::ser::SerializeSeq;
+
+    #[derive(Serialize)]
+    struct Action {
+        #[serde(rename = "android:name")]
+        name: String,
+    };
+    let mut seq = serializer.serialize_seq(Some(actions.len()))?;
+    for action in actions {
+        seq.serialize_element(&Action {
+            name: action.clone(),
+        })?;
+    }
+    seq.end()
+}
+
+fn serialize_catergories<S>(
+    categories: &Option<Vec<String>>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    match categories {
+        Some(categories) => {
+            #[derive(Serialize)]
+            struct Category {
+                #[serde(rename = "android:name")]
+                pub name: String,
+            }
+            let mut c = Vec::new();
+            for category in categories {
+                c.push(Category {
+                    name: category.clone(),
+                });
+            }
+            serializer.serialize_some(&c)
+        }
+        None => serializer.serialize_none(),
+    }
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct IntentFilterData {
+    #[serde(rename(serialize = "android:scheme"))]
+    pub scheme: Option<String>,
+    #[serde(rename(serialize = "android:host"))]
+    pub host: Option<String>,
+    #[serde(rename(serialize = "android:port"))]
+    pub port: Option<String>,
+    #[serde(rename(serialize = "android:path"))]
+    pub path: Option<String>,
+    #[serde(rename(serialize = "android:pathPattern"))]
+    pub path_pattern: Option<String>,
+    #[serde(rename(serialize = "android:pathPrefix"))]
+    pub path_prefix: Option<String>,
+    #[serde(rename(serialize = "android:mimeType"))]
+    pub mime_type: Option<String>,
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct MetaData {
+    #[serde(rename(serialize = "android:name"))]
+    pub name: String,
+    #[serde(rename(serialize = "android:value"))]
+    pub value: String,
+}
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
 pub struct Feature {
-    pub name: String,
-    pub required: bool,
+    #[serde(rename(serialize = "android:name"))]
+    pub name: Option<String>,
+    #[serde(rename(serialize = "android:required"))]
+    pub required: Option<bool>,
+    #[serde(rename(serialize = "android:glEsVersion"))]
+    #[serde(serialize_with = "serialize_opengles_version")]
+    pub opengles_version: Option<(u8, u8)>,
 }
 
-impl Feature {
-    pub fn to_string(&self) -> String {
-        format!(
-            r#"<uses-feature android:name="{}" android:required="{}"/>"#,
-            &self.name, self.required,
-        )
+fn serialize_opengles_version<S>(
+    version: &Option<(u8, u8)>,
+    serializer: S,
+) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    match version {
+        Some(version) => {
+            let opengles_version = format!("0x{:04}{:04}", version.0, version.1);
+            serializer.serialize_some(&opengles_version)
+        }
+        None => serializer.serialize_none(),
     }
 }
 
-#[derive(Debug)]
-pub struct ApplicationMetadata {
-    pub name: String,
-    pub value: String,
-}
-
-impl ApplicationMetadata {
-    pub fn to_string(&self) -> String {
-        format!(
-            r#"<meta-data android:name="{}" android:value="{}"/>"#,
-            self.name, self.value
-        )
-    }
-}
-
-#[derive(Debug)]
-pub struct ActivityMetadata {
-    pub name: String,
-    pub value: String,
-}
-
-impl ActivityMetadata {
-    pub fn to_string(&self) -> String {
-        format!(
-            r#"<meta-data android:name="{}" android:value="{}"/>"#,
-            self.name, self.value
-        )
-    }
-}
-
-#[derive(Debug)]
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
 pub struct Permission {
+    #[serde(rename(serialize = "android:name"))]
     pub name: String,
+    #[serde(rename(serialize = "android:maxSdkVersion"))]
     pub max_sdk_version: Option<u32>,
 }
 
-impl Permission {
-    pub fn to_string(&self) -> String {
-        let max_sdk_version = self
-            .max_sdk_version
-            .as_ref()
-            .map(|max_sdk_version| format!(r#"android:maxSdkVersion="{}""#, max_sdk_version))
-            .unwrap_or_default();
-        format!(
-            r#"<uses-permission android:name="{}" {}/>"#,
-            &self.name, max_sdk_version,
-        )
-    }
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct Sdk {
+    #[serde(rename(serialize = "android:minSdkVersion"))]
+    pub min_sdk_version: Option<u32>,
+    #[serde(rename(serialize = "android:targetSdkVersion"))]
+    pub target_sdk_version: Option<u32>,
+    #[serde(rename(serialize = "android:maxSdkVersion"))]
+    pub max_sdk_version: Option<u32>,
 }
 
-#[derive(Debug)]
-pub struct IntentFilterData {
-    pub scheme: Option<String>,
-    pub host: Option<String>,
-    pub prefix: Option<String>,
-}
-
-impl IntentFilterData {
-    pub fn to_string(&self) -> String {
-        let host = if let Some(host) = self.host.as_ref() {
-            format!(" android:host=\"{}\"", host)
-        } else {
-            "".into()
-        };
-
-        let prefix = if let Some(prefix) = self.prefix.as_ref() {
-            format!(" android:pathPrefix=\"{}\"", prefix)
-        } else {
-            "".into()
-        };
-
-        let scheme = if let Some(scheme) = self.scheme.as_ref() {
-            format!(" android:scheme=\"{}\"", scheme)
-        } else {
-            "".into()
-        };
-
-        format!("<data {} {} {}/>", scheme, &host, &prefix)
-    }
-}
-
-#[derive(Debug)]
-pub struct IntentFilter {
-    pub name: String,
-    pub categories: Vec<String>,
-    pub data: Vec<IntentFilterData>,
-}
-
-impl IntentFilter {
-    pub fn to_string(&self) -> String {
-        let mut categories = "".to_string();
-        for category in &self.categories {
-            categories = format!("{}<category android:name=\"{}\"/>", categories, category)
+impl Default for Sdk {
+    fn default() -> Self {
+        Self {
+            min_sdk_version: Some(23),
+            target_sdk_version: None,
+            max_sdk_version: None,
         }
-
-        let mut data = "".to_string();
-        for d in &self.data {
-            data = format!("{}{}", data, d.to_string())
-        }
-
-        format!(
-            "<intent-filter>
-            \t{}
-            \t{}
-            \t<action android:name=\"{}\"/>
-            </intent-filter>",
-            &categories, &data, &self.name,
-        )
     }
+}
+
+// Manifest namespace attribute should always be set to this value.
+//
+// https://developer.android.com/guide/topics/manifest/manifest-element
+fn default_android_namespace() -> String {
+    "http://schemas.android.com/apk/res/android".to_string()
+}
+
+fn default_has_code() -> Option<bool> {
+    Some(false)
+}
+
+fn default_config_changes() -> Option<String> {
+    Some("orientation|keyboardHidden|screenSize".to_string())
+}
+
+fn default_activity_name() -> Option<String> {
+    Some("android.app.NativeActivity".to_string())
 }